requirements.txt
absl-py==2.1.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.10.0
defusedxml==0.7.1
dm-tree==0.1.8
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h5py==3.10.0
idna==2.10
Jinja2==3.1.3
keras==3.0.5
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.21
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
supervision==0.18.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.23
urllib3==2.2.1
